% File Quor/man/gleason7.Rd
% Part of the Quor package for R (http://www.R-project.org)
% Copyright (C) 2014, Adriano Polpo, Carlos A. de B. Pereira, Cassio P. de Campos.
% Distributed under FDL 1.3 or later
% http://www.gnu.org/licenses/fdl-1.3.html

\name{gleason7}

%%%
\alias{gleason7}
\title{Data set Gleason 7}
\description{
Pre-operative Gleason score provide valuable prognosis in cases of prostate
cancer, in general. However, for patient's Gleason 7 it does not. This is
because Gleason 7 tumors display great morphological heterogeneity among
regions. The data set have the microarray data of gene RPS28 for recurrent
(R) and non-recurrent (NR) Gleason 7 prostate cancer patients.
}
\value{
  The data variables are:
  \item{ R}{Recurrent cases.}
  \item{NR}{Non-recurrent cases.}
}
